Revert "Add files via upload"
This reverts commit 5ca62d5557ea8855c2ad227d01b7852802e9d0a4.

+++ LOOTBOX/Worker.cs
@@ -3,14 +3,14 @@
     public class Worker : Person
     {
         public BoxSize BaseSize { get; }
-        public float Speed { get; }
+        public int Speed { get; }
         public bool IsBusy { get; set; }
         public bool IsFired { get; set; }
         public Worker(string fName,
                       string lName,
                       string mName = "",
                       BoxSize size = BoxSize.Small,
-                      float speed = 5f,
+                      int speed = 5,
                       bool isBusy = false,
                       bool isFired = false) : base(fName, lName, mName)
         {
